- Imports changed to refer the new module - swedish filter to chunker - modifications on owneryear and owner name extraction so that they work with Suuret maatilat
