+++ dataprocessing/books/greatfarmers/extraction/extractors/farmextractor.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
-from books.farmers.extraction.extractors.baseExtractor import BaseExtractor
-from books.farmers.extractionkeys import KEYS
+from books.greatfarmers.extraction.extractors.baseExtractor import BaseExtractor
+from books.greatfarmers.extractionkeys import KEYS
 from interface.valuewrapper import ValueWrapper
-#from books.farmers.extraction.extractionExceptions import ProfessionException
+#from books.greatfarmers.extraction.extractionExceptions import ProfessionException
 import shared.textUtils as textUtils
 import shared.regexUtils as regexUtils
 import re
