+++ dataprocessing/books/greatfarmers/extraction/extractors/hostessextractor.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
-from books.farmers.extraction.extractors.baseExtractor import BaseExtractor
-from books.farmers.extractionkeys import KEYS
+from books.greatfarmers.extraction.extractors.baseExtractor import BaseExtractor
+from books.greatfarmers.extractionkeys import KEYS
 from interface.valuewrapper import ValueWrapper
-from books.farmers.extraction.extractionExceptions import  HostessNameException
-from books.farmers.extraction.extractors.birthdayExtractor import BirthdayExtractor
+from books.greatfarmers.extraction.extractionExceptions import  HostessNameException
+from books.greatfarmers.extraction.extractors.birthdayExtractor import BirthdayExtractor
 import shared.textUtils as textUtils
 import shared.regexUtils as regexUtils
 from shared.genderExtract import Gender
