+++ dataprocessing/books/greatfarmers/extraction/extractors/quantityextractor.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
-from books.farmers.extraction.extractors.baseExtractor import BaseExtractor
-from books.farmers.extractionkeys import KEYS
+from books.greatfarmers.extraction.extractors.baseExtractor import BaseExtractor
+from books.greatfarmers.extractionkeys import KEYS
 from interface.valuewrapper import ValueWrapper
 import shared.textUtils as textUtils
 import shared.regexUtils as regexUtils
