+++ dataprocessing/books/greatfarmers/resultcsvbuilder.py
@@ -3,7 +3,7 @@ import ntpath
 import csv
 import json
 from abc import abstractmethod
-from books.farmers.extractionkeys import KEYS
+from books.greatfarmers.extractionkeys import KEYS
 from interface.valuewrapper import ValueWrapper
 from interface.csvbuilderinterface import ResultCsvBuilderInterface
