+++ dataprocessing/qtgui/xmlImport.py
@@ -101,7 +101,6 @@ class XmlImport(QObject):
 
 
     def _getXMLroot(self, filepath):
-
         #read the data in XML-format to be processed
         parser = etree.XMLParser(encoding="utf-8")
         tree = etree.parse(filepath, parser=parser) #ET.parse(filepath)
